+++ src/GitHub/Data/PullRequests.hs
@@ -14,14 +14,12 @@ module GitHub.Data.PullRequests (
     PullRequestEventType(..),
     PullRequestReference(..),
     MergeResult(..),
-    statusMerge,
     ) where
 
 import GitHub.Data.Definitions
 import GitHub.Data.Id          (Id)
 import GitHub.Data.Options     (IssueState (..), MergeableState (..))
 import GitHub.Data.Repos       (Repo)
-import GitHub.Data.Request     (StatusMap)
 import GitHub.Data.URL         (URL)
 import GitHub.Internal.Prelude
 import Prelude ()
@@ -34,7 +32,7 @@ data SimplePullRequest = SimplePullRequest
     , simplePullRequestUser               :: !SimpleUser
     , simplePullRequestPatchUrl           :: !URL
     , simplePullRequestState              :: !IssueState
-    , simplePullRequestNumber             :: !Int
+    , simplePullRequestNumber             :: !IssueNumber
     , simplePullRequestHtmlUrl            :: !URL
     , simplePullRequestUpdatedAt          :: !UTCTime
     , simplePullRequestBody               :: !(Maybe Text)
@@ -59,7 +57,7 @@ data PullRequest = PullRequest
     , pullRequestUser                 :: !SimpleUser
     , pullRequestPatchUrl             :: !URL
     , pullRequestState                :: !IssueState
-    , pullRequestNumber               :: !Int
+    , pullRequestNumber               :: !IssueNumber
     , pullRequestHtmlUrl              :: !URL
     , pullRequestUpdatedAt            :: !UTCTime
     , pullRequestBody                 :: !(Maybe Text)
@@ -319,10 +317,3 @@ data MergeResult
     | MergeCannotPerform
     | MergeConflict
   deriving (Eq, Ord, Read, Show, Enum, Bounded, Generic, Typeable)
-
-statusMerge :: StatusMap MergeResult
-statusMerge =
-    [ (200, MergeSuccessful)
-    , (405, MergeCannotPerform)
-    , (409, MergeConflict)
-    ]