+++ src/GitHub/Data/Request.hs
@@ -10,36 +10,33 @@
 --
 module GitHub.Data.Request (
     -- * Request
-    Request (..),
-    SimpleRequest (..),
+    Request,
+    GenRequest (..),
     -- * Smart constructors
     query, pagedQuery, command,
     -- * Auxiliary types
     RW(..),
-    StatusMap,
-    statusOnlyOk,
     CommandMethod(..),
     toMethod,
     FetchCount(..),
+    MediaType (..),
     Paths,
     IsPathPart(..),
     QueryString,
     Count,
     ) where
 
-import GitHub.Data.Definitions (Count, QueryString)
+import GitHub.Data.Definitions (Count, IssueNumber, QueryString, unIssueNumber)
 import GitHub.Data.Id          (Id, untagId)
 import GitHub.Data.Name        (Name, untagName)
 import GitHub.Internal.Prelude
 
 import qualified Data.ByteString.Lazy      as LBS
 import qualified Data.Text                 as T
-import qualified Network.HTTP.Types        as Types
 import qualified Network.HTTP.Types.Method as Method
-import Network.URI                         (URI)
 
 ------------------------------------------------------------------------------
--- Auxillary types
+-- Path parts
 ------------------------------------------------------------------------------
 
 type Paths = [Text]
@@ -53,42 +50,32 @@ instance IsPathPart (Name a) where
 instance IsPathPart (Id a) where
     toPathPart = T.pack . show . untagId
 
+instance IsPathPart IssueNumber where
+    toPathPart = T.pack . show . unIssueNumber
+
+-------------------------------------------------------------------------------
+-- Command Method
+-------------------------------------------------------------------------------
+
 -- | Http method of requests with body.
-data CommandMethod a where
-    Post   :: CommandMethod a
-    Patch  :: CommandMethod a
-    Patch' :: CommandMethod ()
-    Put    :: CommandMethod a
-    Put'   :: CommandMethod ()
-    Delete :: CommandMethod ()
-    deriving (Typeable)
+data CommandMethod
+    = Post
+    | Patch
+    | Put
+    | Delete
+  deriving (Eq, Ord, Read, Show, Enum, Bounded, Typeable, Data, Generic)
 
-deriving instance Eq (CommandMethod a)
-deriving instance Ord (CommandMethod a)
+instance Hashable CommandMethod
 
-instance Show (CommandMethod a) where
-    showsPrec _ Post    = showString "Post"
-    showsPrec _ Patch   = showString "Patch"
-    showsPrec _ Patch'  = showString "Patch'"
-    showsPrec _ Put     = showString "Put"
-    showsPrec _ Put'    = showString "Put'"
-    showsPrec _ Delete  = showString "Delete"
-
-instance Hashable (CommandMethod a) where
-    hashWithSalt salt Post    = hashWithSalt salt (0 :: Int)
-    hashWithSalt salt Patch   = hashWithSalt salt (1 :: Int)
-    hashWithSalt salt Put     = hashWithSalt salt (2 :: Int)
-    hashWithSalt salt Put'    = hashWithSalt salt (3 :: Int)
-    hashWithSalt salt Delete  = hashWithSalt salt (4 :: Int)
-    hashWithSalt salt Patch'  = hashWithSalt salt (5 :: Int)
-
-toMethod :: CommandMethod a -> Method.Method
+toMethod :: CommandMethod -> Method.Method
 toMethod Post   = Method.methodPost
 toMethod Patch  = Method.methodPatch
-toMethod Patch' = Method.methodPatch
 toMethod Put    = Method.methodPut
-toMethod Put'   = Method.methodPut
 toMethod Delete = Method.methodDelete
+
+-------------------------------------------------------------------------------
+-- Fetch count
+-------------------------------------------------------------------------------
 
 -- | 'PagedQuery' returns just some results, using this data we can specify how
 -- many pages we want to fetch.
@@ -114,15 +101,31 @@ instance Hashable FetchCount
 instance Binary FetchCount
 instance NFData FetchCount where rnf = genericRnf
 
+-------------------------------------------------------------------------------
+-- MediaType
+-------------------------------------------------------------------------------
+
+data MediaType
+    = MtJSON     -- ^ @application/vnd.github.v3+json@
+    | MtRaw      -- ^ @application/vnd.github.v3.raw@ <https://developer.github.com/v3/media/#raw-1>
+    | MtDiff     -- ^ @application/vnd.github.v3.diff@ <https://developer.github.com/v3/media/#diff>
+    | MtPatch    -- ^ @application/vnd.github.v3.patch@ <https://developer.github.com/v3/media/#patch>
+    | MtSha      -- ^ @application/vnd.github.v3.sha@ <https://developer.github.com/v3/media/#sha>
+    | MtStar     -- ^ @application/vnd.github.v3.star+json@ <https://developer.github.com/v3/activity/starring/#alternative-response-with-star-creation-timestamps-1>
+    | MtRedirect -- ^ <https://developer.github.com/v3/repos/contents/#get-archive-link>
+    | MtStatus   -- ^ Parse status
+    | MtUnit     -- ^ Always succeeds
+  deriving (Eq, Ord, Read, Show, Enum, Bounded, Typeable, Data, Generic)
+
 ------------------------------------------------------------------------------
--- Github request
+-- RW
 ------------------------------------------------------------------------------
 
 -- | Type used as with @DataKinds@ to tag whether requests need authentication
 -- or aren't read-only.
 data RW
     = RO  -- ^ /Read-only/, doesn't necessarily requires authentication
-    | RA  -- ^ /Read autenticated/
+    | RA  -- ^ /Read authenticated/
     | RW  -- ^ /Read-write/, requires authentication
   deriving (Eq, Ord, Read, Show, Enum, Bounded, Typeable, Data, Generic)
 
@@ -137,100 +140,54 @@ instance IReadOnly 'RO        where iro = ROO
 instance IReadOnly 'RA        where iro = ROA
 -}
 
+-------------------------------------------------------------------------------
+-- GitHub Request
+-------------------------------------------------------------------------------
+
 -- | Github request data type.
 --
--- * @k@ describes whether authentication is required. It's required for non-@GET@ requests.
+-- * @rw@ describes whether authentication is required. It's required for non-@GET@ requests.
+-- * @mt@ describes the media type, i.e. how the response should be interpreted.
 -- * @a@ is the result type
 --
 -- /Note:/ 'Request' is not 'Functor' on purpose.
-data Request (k :: RW) a where
-    SimpleQuery   :: FromJSON a => SimpleRequest k a -> Request k a
-    StatusQuery   :: StatusMap a -> SimpleRequest k () -> Request k a
-    HeaderQuery   :: FromJSON a => Types.RequestHeaders -> SimpleRequest k a -> Request k a
-    RedirectQuery :: SimpleRequest k () -> Request k URI
+data GenRequest (mt :: MediaType) (rw :: RW) a where
+    Query        :: Paths -> QueryString -> GenRequest mt rw a
+    PagedQuery   :: Paths -> QueryString -> FetchCount -> GenRequest mt rw (Vector a)
+
+    -- | Command
+    Command
+        :: CommandMethod           -- ^ command
+        -> Paths                   -- ^ path
+        -> LBS.ByteString          -- ^ body
+        -> GenRequest mt 'RW a
   deriving (Typeable)
 
-data SimpleRequest (k :: RW) a where
-    Query        :: Paths -> QueryString -> SimpleRequest k a
-    PagedQuery   :: Paths -> QueryString -> FetchCount -> SimpleRequest k (Vector a)
-    Command      :: CommandMethod a -> Paths -> LBS.ByteString -> SimpleRequest 'RW a
-  deriving (Typeable)
-
--------------------------------------------------------------------------------
--- Status Map
--------------------------------------------------------------------------------
-
--- TODO: Change to 'Map' ?
-type StatusMap a = [(Int, a)]
-
-statusOnlyOk :: StatusMap Bool
-statusOnlyOk =
-    [ (204, True)
-    , (404, False)
-    ]
+-- | Most requests ask for @JSON@.
+type Request = GenRequest 'MtJSON
 
 -------------------------------------------------------------------------------
 -- Smart constructors
 -------------------------------------------------------------------------------
 
-query :: FromJSON a => Paths -> QueryString -> Request k a
-query ps qs = SimpleQuery (Query ps qs)
+query :: Paths -> QueryString -> Request mt a
+query ps qs = Query ps qs
 
-pagedQuery :: FromJSON a => Paths -> QueryString -> FetchCount -> Request k (Vector a)
-pagedQuery ps qs fc = SimpleQuery (PagedQuery ps qs fc)
+pagedQuery :: FromJSON a => Paths -> QueryString -> FetchCount -> Request mt (Vector a)
+pagedQuery ps qs fc = PagedQuery ps qs fc
 
-command :: FromJSON a => CommandMethod a -> Paths -> LBS.ByteString -> Request 'RW a
-command m ps body = SimpleQuery (Command m ps body)
+command :: CommandMethod -> Paths -> LBS.ByteString -> Request 'RW a
+command m ps body = Command m ps body
 
 -------------------------------------------------------------------------------
 -- Instances
 -------------------------------------------------------------------------------
 
-deriving instance Eq a => Eq (Request k a)
-deriving instance Eq a => Eq (SimpleRequest k a)
+deriving instance Eq (GenRequest rw mt a)
+deriving instance Ord (GenRequest rw mt a)
+deriving instance Show (GenRequest rw mt a)
 
-deriving instance Ord a => Ord (Request k a)
-deriving instance Ord a => Ord (SimpleRequest k a)
-
-instance Show (SimpleRequest k a) where
-    showsPrec d r = showParen (d > appPrec) $ case r of
-        Query ps qs -> showString "Query "
-            . showsPrec (appPrec + 1) ps
-            . showString " "
-            . showsPrec (appPrec + 1) qs
-        PagedQuery ps qs l -> showString "PagedQuery "
-            . showsPrec (appPrec + 1) ps
-            . showString " "
-            . showsPrec (appPrec + 1) qs
-            . showString " "
-            . showsPrec (appPrec + 1) l
-        Command m ps body -> showString "Command "
-            . showsPrec (appPrec + 1) m
-            . showString " "
-            . showsPrec (appPrec + 1) ps
-            . showString " "
-            . showsPrec (appPrec + 1) body
-      where
-        appPrec = 10 :: Int
-
-instance Show (Request k a) where
-    showsPrec d r = showParen (d > appPrec) $ case r of
-        SimpleQuery req -> showString "SimpleQuery "
-            . showsPrec (appPrec + 1) req
-        StatusQuery m req -> showString "Status "
-            . showsPrec (appPrec + 1) (map fst m) -- !!! printing only keys
-            . showString " "
-            . showsPrec (appPrec + 1) req
-        HeaderQuery m req -> showString "Header "
-            . showsPrec (appPrec + 1) m
-            . showString " "
-            . showsPrec (appPrec + 1) req
-        RedirectQuery req -> showString "Redirect "
-            . showsPrec (appPrec + 1) req
-      where
-        appPrec = 10 :: Int
-
-instance Hashable (SimpleRequest k a) where
+instance Hashable (GenRequest rw mt a) where
     hashWithSalt salt (Query ps qs) =
         salt `hashWithSalt` (0 :: Int)
              `hashWithSalt` ps
@@ -246,18 +203,4 @@ instance Hashable (SimpleRequest k a) where
              `hashWithSalt` ps
              `hashWithSalt` body
 
-instance Hashable (Request k a) where
-    hashWithSalt salt (SimpleQuery req) =
-        salt `hashWithSalt` (0 :: Int)
-             `hashWithSalt` req
-    hashWithSalt salt (StatusQuery sm req) =
-        salt `hashWithSalt` (1 :: Int)
-             `hashWithSalt` map fst sm
-             `hashWithSalt` req
-    hashWithSalt salt (HeaderQuery h req) =
-        salt `hashWithSalt` (2 :: Int)
-             `hashWithSalt` h
-             `hashWithSalt` req
-    hashWithSalt salt (RedirectQuery req) =
-        salt `hashWithSalt` (3 :: Int)
-             `hashWithSalt` req
+-- TODO: Binary