+++ src/GitHub/Endpoints/Gists.hs
@@ -27,7 +27,7 @@ import Prelude ()
 
 -- | The list of all gists created by the user
 --
--- > gists' (Just ("github-username", "github-password")) "mike-burns"
+-- > gists' (Just $ BasicAuth "github-username" "github-password") "mike-burns"
 gists' :: Maybe Auth -> Name Owner -> IO (Either Error (Vector Gist))
 gists' auth user =
     executeRequestMaybe auth $ gistsR user FetchAll
@@ -45,7 +45,7 @@ gistsR user = pagedQuery ["users", toPathPart user, "gists"] []
 
 -- | A specific gist, given its id, with authentication credentials
 --
--- > gist' (Just ("github-username", "github-password")) "225074"
+-- > gist' (Just $ BasicAuth "github-username" "github-password") "225074"
 gist' :: Maybe Auth -> Name Gist -> IO (Either Error Gist)
 gist' auth gid =
     executeRequestMaybe auth $ gistR gid
@@ -64,18 +64,18 @@ gistR gid =
 
 -- | Star a gist by the authenticated user.
 --
--- > starGist ("github-username", "github-password") "225074"
+-- > starGist (BasicAuth "github-username" "github-password") "225074"
 starGist :: Auth -> Name Gist -> IO (Either Error ())
 starGist auth gid = executeRequest auth $ starGistR gid
 
 -- | Star a gist by the authenticated user.
 -- See <https://developer.github.com/v3/gists/#star-a-gist>
-starGistR :: Name Gist -> Request 'RW ()
-starGistR gid = command Put' ["gists", toPathPart gid, "star"] mempty
+starGistR :: Name Gist -> GenRequest 'MtUnit 'RW ()
+starGistR gid = Command Put ["gists", toPathPart gid, "star"] mempty
 
 -- | Unstar a gist by the authenticated user.
 --
--- > unstarGist ("github-username", "github-password") "225074"
+-- > unstarGist (BasicAuth "github-username" "github-password") "225074"
 unstarGist :: Auth -> Name Gist -> IO (Either Error ())
 unstarGist auth gid = executeRequest auth $ unstarGistR gid
 
@@ -86,7 +86,7 @@ unstarGistR gid = command Delete ["gists", toPathPart gid, "star"] mempty
 
 -- | Delete a gist by the authenticated user.
 --
--- > deleteGist ("github-username", "github-password") "225074"
+-- > deleteGist (BasicAuth "github-username" "github-password") "225074"
 deleteGist :: Auth -> Name Gist -> IO (Either Error ())
 deleteGist auth gid = executeRequest auth $ deleteGistR gid