+++ src/GitHub/Endpoints/GitData/Blobs.hs
@@ -19,7 +19,7 @@ import Prelude ()
 
 -- | Query a blob by SHA1.
 --
--- > blob' (Just ("github-username", "github-password")) "thoughtbot" "paperclip" "bc5c51d1ece1ee45f94b056a0f5a1674d7e8cba9"
+-- > blob' (Just $ BasicAuth "github-username" "github-password") "thoughtbot" "paperclip" "bc5c51d1ece1ee45f94b056a0f5a1674d7e8cba9"
 blob' :: Maybe Auth -> Name Owner -> Name Repo -> Name Blob -> IO (Either Error Blob)
 blob' auth user repo sha =
     executeRequestMaybe auth $ blobR user repo sha