+++ src/GitHub/Endpoints/Issues/Milestones.hs
@@ -13,6 +13,10 @@ module GitHub.Endpoints.Issues.Milestones (
     milestoneR,
     createMilestone,
     createMilestoneR,
+    updateMilestone,
+    updateMilestoneR,
+    deleteMilestone,
+    deleteMilestoneR,
     module GitHub.Data,
     ) where
 
@@ -29,7 +33,7 @@ milestones = milestones' Nothing
 
 -- | All milestones in the repo, using authentication.
 --
--- > milestones' (User (user, passwordG) "thoughtbot" "paperclip"
+-- > milestones' (Just $ BasicAuth "github-username" "github-password") "thoughtbot" "paperclip"
 milestones' :: Maybe Auth -> Name Owner -> Name Repo -> IO (Either Error (Vector Milestone))
 milestones' auth user repo =
     executeRequestMaybe auth $ milestonesR user repo FetchAll
@@ -54,7 +58,7 @@ milestoneR user repo mid =
     query ["repos", toPathPart user, toPathPart repo, "milestones", toPathPart mid] []
 
 createMilestone :: Auth -> Name Owner -> Name Repo -> NewMilestone -> IO (Either Error Milestone)
-createMilestone auth user repo mst = executeRequest auth $ createMilestoneR user repo mst
+createMilestone auth user repo mlstn = executeRequest auth $ createMilestoneR user repo mlstn
 
 -- | Create a milestone.
 -- See <https://developer.github.com/v3/issues/milestones/#create-a-milestone>
@@ -62,3 +66,21 @@ createMilestoneR :: Name Owner -> Name Repo -> NewMilestone -> Request 'RW Miles
 createMilestoneR user repo =
     command Post ["repos", toPathPart user, toPathPart repo, "milestones"] . encode
 
+updateMilestone :: Auth -> Name Owner -> Name Repo -> Id Milestone -> UpdateMilestone -> IO (Either Error Milestone)
+updateMilestone auth user repo mid mlstn = executeRequest auth $ updateMilestoneR user repo mid mlstn
+
+-- | Update a milestone.
+-- See <https://developer.github.com/v3/issues/milestones/#update-a-milestone>
+updateMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> UpdateMilestone -> Request 'RW Milestone
+updateMilestoneR user repo mid =
+    command Patch ["repos", toPathPart user, toPathPart repo, "milestones", toPathPart mid ] . encode
+
+deleteMilestone :: Auth -> Name Owner -> Name Repo -> Id Milestone -> IO (Either Error ())
+deleteMilestone auth user repo mid = executeRequest auth $ deleteMilestoneR user repo mid
+
+-- | Delete a milestone.
+-- See <https://developer.github.com/v3/issues/milestones/#delete-a-milestone>
+deleteMilestoneR :: Name Owner -> Name Repo -> Id Milestone -> Request 'RW ()
+deleteMilestoneR user repo mid =
+    command Delete
+        ["repos", toPathPart user, toPathPart repo, "milestones", toPathPart mid] mempty