+++ src/GitHub/Endpoints/Organizations.hs
@@ -11,6 +11,7 @@ module GitHub.Endpoints.Organizations (
     publicOrganization,
     publicOrganization',
     publicOrganizationR,
+    organizationsR,
     module GitHub.Data,
     ) where
 
@@ -21,7 +22,7 @@ import Prelude ()
 
 -- | The public organizations for a user, given the user's login, with authorization
 --
--- > publicOrganizationsFor' (Just ("github-username", "github-password")) "mike-burns"
+-- > publicOrganizationsFor' (Just $ BasicAuth "github-username" "github-password") "mike-burns"
 publicOrganizationsFor' :: Maybe Auth -> Name User -> IO (Either Error (Vector SimpleOrganization))
 publicOrganizationsFor' auth org =
     executeRequestMaybe auth $ publicOrganizationsForR org FetchAll
@@ -32,14 +33,19 @@ publicOrganizationsFor' auth org =
 publicOrganizationsFor :: Name User -> IO (Either Error (Vector SimpleOrganization))
 publicOrganizationsFor = publicOrganizationsFor' Nothing
 
--- | List user organizations.
+-- | List all user organizations.
+-- See <https://developer.github.com/v3/orgs/#list-your-organizations>
+organizationsR :: FetchCount -> Request k (Vector SimpleOrganization)
+organizationsR = pagedQuery ["user", "orgs"] []
+
+-- | List public user organizations.
 -- See <https://developer.github.com/v3/orgs/#list-user-organizations>
 publicOrganizationsForR :: Name User -> FetchCount -> Request k (Vector SimpleOrganization)
 publicOrganizationsForR user = pagedQuery ["users", toPathPart user, "orgs"] []
 
 -- | Details on a public organization. Takes the organization's login.
 --
--- > publicOrganization' (Just ("github-username", "github-password")) "thoughtbot"
+-- > publicOrganization' (Just $ BasicAuth "github-username" "github-password") "thoughtbot"
 publicOrganization' :: Maybe Auth -> Name Organization -> IO (Either Error Organization)
 publicOrganization' auth = executeRequestMaybe auth . publicOrganizationR