+++ src/GitHub/Internal/Prelude.hs
@@ -4,7 +4,7 @@
 -- License     :  BSD-3-Clause
 -- Maintainer  :  Oleg Grenrus <oleg.grenrus@iki.fi>
 --
--- This module may change between minor releases. Do not rely on it contents.
+-- This module may change between minor releases. Do not rely on its contents.
 module GitHub.Internal.Prelude (
     module Prelude.Compat,
     -- * Commonly used types